array implemeted

+++ src/parser/Parser.java
@@ -6,5 +6,5 @@ public interface Parser {
 
     boolean parse(Grammar g, String s);
 
-    int getCounter();
+    long getCounter();
 }
